show images anyway

+++ app/src/main/java/com/example/cs496_proj1/Gallery/ImageUnit.java
@@ -3,10 +3,10 @@ package com.example.cs496_proj1.Gallery;
 import android.net.Uri;
 
 public class ImageUnit {
-    int id;
+    long id;
     Uri imageUri;
 
-    public ImageUnit(int id, Uri imageUri) {
+    public ImageUnit(long id, Uri imageUri) {
         super();
         this.id = id;
         this.imageUri = imageUri;
